# sso_manual/openapi.json
{
  "swagger": "2.0",
  "definitions": {
...
    "RequestAuditCreate": {
      "type": "object",
      "properties": {
        "api_key_id": {
          "type": "string",
          "format": "uuid"
        },
        "audit_type": {
          "type": "string"
        },
        "data": {
          "description": "Audit log data object",
          "type": "object"
        },
        "status_code": {
          "type": "integer",
          "format": "int32"
        },
        "subject": {
          "description": "Audit log subject",
          "type": "string"
        },
        "token_id": {
          "type": "string",
          "format": "uuid"
        },
        "user_id": {
          "description": "Audit log user UUID",
          "type": "string",
          "format": "uuid"
        }
      },
      "required": [
        "audit_type",
        "data"
      ]
    },
...
    "RequestCsrf": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "required": [
        "token"
      ]
    },
...
    "ResponseAudit": {
      "type": "object",
      "properties": {
        "api_key_id": {
          "type": "string",
          "format": "uuid"
        },
        "audit_type": {
          "type": "string"
        },
        "client_id": {
          "type": "string",
          "format": "uuid"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "data": {
          "type": "object"
        },
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "status_code": {
          "type": "integer",
          "format": "int32"
        },
        "subject": {
          "type": "string"
        },
        "token_id": {
          "type": "string",
          "format": "uuid"
        },
        "user_id": {
          "type": "string",
          "format": "uuid"
        }
      },
      "required": [
        "audit_type",
        "created_at",
        "id"
      ]
    },
    "ResponseAuditMany": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "api_key_id": {
                "type": "string",
                "format": "uuid"
              },
              "audit_type": {
                "type": "string"
              },
              "client_id": {
                "type": "string",
                "format": "uuid"
              },
              "created_at": {
                "type": "string",
                "format": "date-time"
              },
              "data": {
                "type": "object"
              },
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "status_code": {
                "type": "integer",
                "format": "int32"
              },
              "subject": {
                "type": "string"
              },
              "token_id": {
                "type": "string",
                "format": "uuid"
              },
              "user_id": {
                "type": "string",
                "format": "uuid"
              }
            },
            "required": [
              "audit_type",
              "created_at",
              "id"
            ]
          }
        }
      },
      "required": [
        "data"
      ]
    },
...
    "ResponseCsrf": {
      "type": "object",
      "properties": {
        "client_id": {
          "type": "string",
          "format": "uuid"
        },
        "created_at": {
          "type": "string",
          "format": "date-time"
        },
        "token": {
          "type": "string"
        },
        "ttl": {
          "type": "string",
          "format": "date-time"
        }
      },
      "required": [
        "client_id",
        "created_at",
        "token",
        "ttl"
      ]
    },
    "ResponseOpenidConfiguration": {
      "type": "object",
      "properties": {
        "authorization_endpoint": {
          "type": "string"
        },
        "issuer": {
          "type": "string"
        },
        "token_endpoint": {
          "type": "string"
        },
        "token_endpoint_auth_methods_supported": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      },
      "required": [
        "authorization_endpoint",
        "issuer",
        "token_endpoint",
        "token_endpoint_auth_methods_supported"
      ]
    },
...
  },
  "paths": {
    "/.well-known/change-password": {
      "get": {
        "summary": "Change password endpoint",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/.well-known/openid-configuration": {
      "get": {
        "summary": "OpenID connect discovery endpoint",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseOpenidConfiguration"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/script/password_strength.js": {
      "get": {
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/script/zxcvbn.js": {
      "get": {
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v2/audit/create": {
      "post": {
        "summary": "Create audit log",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseAudit"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestAuditCreate"
            }
          }
        ]
      }
    },
    "/v2/audit/read": {
      "post": {
        "summary": "Read audit logs",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseAuditMany"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestAuditRead"
            }
          }
        ]
      }
    },
    "/v2/auth/delete": {
      "post": {
        "summary": "Delete user interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "code",
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/v2/auth/email-update": {
      "post": {
        "summary": "User update email interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "message",
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "email_confirm",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "email_new",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/v2/auth/logout": {
      "get": {
        "summary": "User logout interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "message",
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/v2/auth/password-reset": {
      "post": {
        "summary": "User reset password interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_confirm",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_new",
            "type": "string"
          }
        ]
      }
    },
    "/v2/auth/password-update": {
      "post": {
        "summary": "User update password interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "message",
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_confirm",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_new",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/v2/auth/register": {
      "post": {
        "summary": "User registration interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "oauth2_provider",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_allow_reset",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_confirm",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "register_type",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/v2/client/read": {
      "post": {
        "summary": "Read authenticated client",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseClient"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v2/csrf/create": {
      "post": {
        "summary": "Create CSRF token",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseCsrf"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v2/csrf/verify": {
      "post": {
        "summary": "Verify CSRF token",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {}
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestCsrf"
            }
          }
        ]
      }
    },
    "/v2/oauth2/authorize": {
      "post": {
        "summary": "OAuth2 authorization endpoint",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "scope",
            "type": "string"
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "auth_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "oauth2_provider",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "type": "string"
          }
        ]
      }
    },
    "/v2/oauth2/introspect": {
      "post": {
        "summary": "OAuth2 token introspection endpoint",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestOauth2Introspect"
            }
          }
        ]
      }
    },
    "/v2/oauth2/redirect": {
      "get": {
        "summary": "OAuth2 redirect endpoint",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/v2/oauth2/token": {
      "post": {
        "summary": "OAuth2 token endpoint",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestOauth2Token"
            }
          }
        ]
      }
    },
...
    "/v2/user/create": {
      "post": {
        "summary": "Create user",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseUser"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestUserCreate"
            }
          }
        ]
      }
    },
    "/v2/user/delete": {
      "post": {
        "summary": "Delete user",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {}
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestUserDelete"
            }
          }
        ]
      }
    },
    "/v2/user/read": {
      "post": {
        "summary": "Read users",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseUserMany"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestUserRead"
            }
          }
        ]
      }
    },
    "/v2/user/update": {
      "post": {
        "summary": "Update user",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseUser"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestUserUpdate"
            }
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  },
  "info": {
    "version": "v2",
    "title": "Single Sign-On API"
  }
}